requirements.txt
# 主要依赖
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langchain-chroma>=0.2.0
langchain-huggingface>=0.1.0
langchain-google-genai>=2.0.0
langchain-ollama>=0.2.0
streamlit>=1.40.0
chromadb>=0.5.0
scikit-learn>=1.4.0
transformers>=4.40.0
duckduckgo-search>=6.0.0
python-dotenv>=1.0.0
numpy>=1.26.0
pandas>=2.2.0
plotly>=5.24.0
pydantic>=2.0.0
pypdf>=5.0.0
